backend/requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.0
passlib==1.7.4
python-jose==3.3.0
email-validator==2.2.0
pydantic-settings==2.6.1
